style: move cardcreatorInitStates
