+++ src/pages/CardCreator/InputComponents/ExpireDatesInputList/ExpireDateInput.stories.tsx
@@ -1,9 +1,11 @@
 import '@/styles/index.css';
 
-import React, { useState } from 'react';
+import React from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 
-import { expireDatesInit } from '../../CardCreatorInits';
+import { ErrorContextProvider } from '@/stores/ErrorContext';
+import { CardInfoProvider } from '@/stores/CardCreatorContext';
+
 import { ExpireDatesInputListPure } from './ExpireDatesInputList';
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
@@ -19,8 +21,13 @@ export default {
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
 const Template: ComponentStory<typeof ExpireDatesInputListPure> = () => {
-  const [expireDates, setExpireDates] = useState(expireDatesInit);
-  return <ExpireDatesInputListPure />;
+  return (
+    <ErrorContextProvider>
+      <CardInfoProvider>
+        <ExpireDatesInputListPure />
+      </CardInfoProvider>
+    </ErrorContextProvider>
+  );
 };
 
 export const Primary = Template.bind({});
